# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Talos API
  description: >
    Source of truth for the talos-api public (/api/v1) and internal (/internal/v1) REST surface.
    The Angular client is generated from this file (openapi-generator, typescript-angular); CI
    diffs the running API's springdoc output against it. Endpoints are added phase by phase per
    Section 10 of the implementation plan — this file currently covers Phase 1 (auth),
    Phase 3 (project registry), Phase 4 (task board), Phase 5 (agent run lifecycle, API side),
    Phase 6 (cancel + internal changes ingest), Phase 8 (review/approval flow: diff serving and
    the approvals endpoints), Phase 9 (git push and PR workflow: integrations, the
    git-token/pull-request internal endpoints, and GET /runs/{id}/pull-request), Phase 10
    (Dokploy integration), Phase 12 (chat trigger support), Phase 13 (memory ingestion and
    internal retrieval), Phase 14 (per-run cost tracking and advisory recommendations), and
    Phase 15 (multi-user RBAC enforcement and user management).
  version: "0.15.0"
servers:
  - url: /api/v1
    description: "Public API (JWT via Authorization: Bearer)"
  - url: /internal/v1
    description: "Internal API (service token via X-Talos-Internal-Token)"
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      operationId: listProjects
      summary: List projects.
      tags: [projects]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProjectStatus"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageProjectSummary"
    post:
      operationId: createProject
      summary: Register a new project.
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /projects/{id}:
    get:
      operationId: getProject
      summary: Get a project, its active config, and its 5 most recent runs.
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          $ref: "#/components/responses/Error"
    put:
      operationId: updateProject
      summary: Update project fields.
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /projects/{id}/sync-config:
    post:
      operationId: syncProjectConfig
      summary: Validate and store a new talos.yaml version; deactivates the previous active version.
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncConfigRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectConfig"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          description: talos.yaml failed schema validation; error.details is a field-path -> message map.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{id}/environments:
    get:
      operationId: listProjectEnvironments
      summary: The project's deploy targets and their most recent Dokploy deployment status (Phase 10).
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectEnvironment"
        "404":
          $ref: "#/components/responses/Error"
...
  /runs:
    get:
      operationId: listRuns
      summary: List agent runs.
      tags: [runs]
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RunStatus"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageRunSummary"
  /runs/{id}:
    get:
      operationId: getRun
      summary: Get a run and its steps.
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunDetail"
        "404":
          $ref: "#/components/responses/Error"
  /runs/{id}/logs:
    get:
      operationId: getRunLogs
      summary: Get a page of a run's logs, optionally after a given sequence number.
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: afterSequence
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageLogEntry"
        "404":
          $ref: "#/components/responses/Error"
  /runs/{id}/events/stream:
    get:
      operationId: streamRunEvents
      summary: Server-Sent Events stream of log/status/step events for a run (Section 10.3).
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: >
            text/event-stream. Named events: log ({sequence, stream, message, timestamp}),
            status ({from, to}), step ({stepType, status}). Heartbeat comment every 15s.
            Reconnect with Last-Event-ID = last log sequence to backfill before live relay resumes.
          content:
            text/event-stream:
              schema:
                type: string
  /runs/{id}/cancel:
    post:
      operationId: cancelRun
      summary: >
        Cancel a non-terminal run (Section 8.2 cancel edge). Transitions the run to CANCELLED
        immediately and publishes run.cancel.requested so the orchestrator can kill the process
        tree for a run it is actively processing (Phase 6 extension of Section 11 -- see
        packages/contracts/events/run.cancel.requested.json).
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          description: Run is not in a cancellable state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /runs/{id}/diff:
    get:
      operationId: getRunDiff
      summary: >
        Get the run's changed-file summary and full unified diff text (Section 8.1 step 8; Phase 8
        Review Center). diff is null until the orchestrator's REVIEWING step has recorded it via
        /internal/v1/runs/{id}/changes.
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diff"
        "404":
          $ref: "#/components/responses/Error"
  /runs/{id}/pull-request:
    get:
      operationId: getRunPullRequest
      summary: >
        The PR opened for this run after approval (Phase 9). 404 until PullRequestService has
        opened one.
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
        "404":
          $ref: "#/components/responses/Error"
  /runs/{id}/deploy:
    post:
      operationId: deployRun
      summary: >
        Request a deploy of a COMPLETED run's code (Phase 10). Creates a PENDING DEPLOY approval
        and does NOT call the deploy provider when approval is required (production always is,
        Section 12.2; staging follows talos.yaml's deploy.approval_required) -- this is the
        server-side enforcement behind "deploy button disabled without approval". Otherwise
        triggers immediately.
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployTriggerResponse"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          description: "Run is not COMPLETED, or the project has no deploy: block configured."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: getRunDeployStatus
      summary: The current deploy target/status for this run's project (re-resolved from talos.yaml each call). 404 if never requested.
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectEnvironment"
        "404":
          $ref: "#/components/responses/Error"
...
  /internal/v1/runs/{id}/logs:
    post:
      operationId: internalIngestRunLogs
      summary: Ingest a batch of log lines; persists then relays to any open SSE connection via Redis.
      tags: [internal]
      security:
        - serviceToken: []
      servers:
        - url: /internal/v1
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InternalLogsRequest"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/Error"
...
  /internal/v1/runs/{id}/pull-request:
    post:
      operationId: internalCreateRunPullRequest
      summary: >
        Called once the runner supervisor has pushed branchName: opens the PR via the GitHub REST
        API, stores the pull_requests row, publishes pr.created, and transitions the run
        APPROVED -> COMPLETED (Section 8.2).
      tags: [internal]
      security:
        - serviceToken: []
      servers:
        - url: /internal/v1
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InternalPullRequestRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          description: Run is not APPROVED.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceToken:
      type: apiKey
      in: header
      name: X-Talos-Internal-Token
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RunId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error envelope (Section 10.1).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: Error envelope returned on every non-2xx response (Section 10.1).
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: SCREAMING_SNAKE error code enumerated in this spec.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    ProjectStatus:
      type: string
      enum: [ACTIVE, ARCHIVED]
    ProjectSummary:
      type: object
      required: [id, name, slug, repoUrl, defaultBranch, stackType, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        repoUrl:
          type: string
        defaultBranch:
          type: string
        stackType:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Project:
      allOf:
        - $ref: "#/components/schemas/ProjectSummary"
    ProjectConfig:
      type: object
      required: [id, projectId, configYaml, parsedJson, version, active, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        configYaml:
          type: string
        parsedJson:
          type: object
          additionalProperties: true
        version:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    RunSummary:
      type: object
      required: [id, taskId, status, agentKey, createdAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        status:
          type: string
        agentKey:
          type: string
        createdAt:
          type: string
          format: date-time
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/ProjectSummary"
        - type: object
          required: [activeConfig, configHistory, recentRuns]
          properties:
            activeConfig:
              oneOf:
                - $ref: "#/components/schemas/ProjectConfig"
                - type: "null"
            configHistory:
              type: array
              description: Every config version for this project, newest first (Phase 3 acceptance -- version history retained and visible).
              items:
                $ref: "#/components/schemas/ProjectConfig"
            recentRuns:
              type: array
              items:
                $ref: "#/components/schemas/RunSummary"
              maxItems: 5
    CreateProjectRequest:
      type: object
      required: [name, repoUrl, stackType]
      properties:
        name:
          type: string
        repoUrl:
          type: string
        defaultBranch:
          type: string
          description: Defaults to "main" when omitted.
        stackType:
          type: string
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        repoUrl:
          type: string
        defaultBranch:
          type: string
        stackType:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
    SyncConfigRequest:
      type: object
      required: [configYaml]
      properties:
        configYaml:
          type: string
...
    PageProjectSummary:
      type: object
      required: [content, page, size, totalElements]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProjectSummary"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
...
    RunStatus:
      type: string
      enum: [CREATED, QUEUED, PREPARING_WORKSPACE, RUNNING_AGENT, RUNNING_TESTS, REVIEWING,
        WAITING_APPROVAL, APPROVED, REJECTED, COMPLETED, FAILED, CANCELLED]
    TestStatus:
      type: string
      enum: [NOT_RUN, PASSED, FAILED, ERROR]
    ReviewStatus:
      type: string
      enum: [CLEAN, RISK_FLAGGED]
    StepType:
      type: string
      enum: [WORKSPACE, AGENT, TESTS, REVIEW, PUSH, PR, DEPLOY]
    StepStatus:
      type: string
      enum: [RUNNING, COMPLETED, FAILED, SKIPPED]
    LogStream:
      type: string
      enum: [STDOUT, STDERR, SYSTEM]
    Run:
      type: object
      required: [id, taskId, projectId, status, agentKey, providerAuthMode, testStatus, reviewStatus,
        createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RunStatus"
        agentKey:
          type: string
        providerAuthMode:
          type: string
        prompt:
          type: string
        branchName:
          type: string
        workspacePath:
          type: string
        summary:
          type: string
        testStatus:
          $ref: "#/components/schemas/TestStatus"
        reviewStatus:
          $ref: "#/components/schemas/ReviewStatus"
        errorMessage:
          type: string
        exitCode:
          type: integer
        inputTokens:
          type: integer
          description: Phase 14 (Section 16) normalized usage metadata; null when the adapter can't report it.
        outputTokens:
          type: integer
        costUsd:
          type: number
          description: >-
            Provider-reported price in USD. Always null for provider_auth_mode=subscription_local
            (Section 13) -- never a fabricated/estimated price.
        costModel:
          type: string
        timeoutAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Step:
      type: object
      required: [id, runId, stepType, status]
      properties:
        id:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        stepType:
          $ref: "#/components/schemas/StepType"
        status:
          $ref: "#/components/schemas/StepStatus"
        summary:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    RunDetail:
      allOf:
        - $ref: "#/components/schemas/Run"
        - type: object
          required: [steps]
          properties:
            steps:
              type: array
              items:
                $ref: "#/components/schemas/Step"
    LogEntry:
      type: object
      required: [sequence, stream, message, createdAt]
      properties:
        sequence:
          type: integer
          format: int64
        stream:
          $ref: "#/components/schemas/LogStream"
        message:
          type: string
        createdAt:
          type: string
          format: date-time
    PageRunSummary:
      type: object
      required: [content, page, size, totalElements]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/RunSummary"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
    PageLogEntry:
      type: object
      required: [content, page, size, totalElements]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LogEntry"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
...
    InternalLogEntry:
      type: object
      required: [stream, sequence, message, timestamp]
      properties:
        stream:
          $ref: "#/components/schemas/LogStream"
        sequence:
          type: integer
          format: int64
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    InternalLogsRequest:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/InternalLogEntry"
    GitChangeType:
      type: string
      enum: [ADDED, MODIFIED, DELETED, RENAMED]
    GitChange:
      type: object
      required: [filePath, changeType, additions, deletions]
      properties:
        filePath:
          type: string
        changeType:
          $ref: "#/components/schemas/GitChangeType"
        additions:
          type: integer
        deletions:
          type: integer
    GitChangeSummary:
      description: >
        Output-only counterpart to GitChange (Phase 8): adds the Section 12.1 post-run policy scan
        result, shown in the Review Center's file list and risk panel.
      allOf:
        - $ref: "#/components/schemas/GitChange"
        - type: object
          required: [riskFlagged]
          properties:
            riskFlagged:
              type: boolean
            matchedPattern:
              type: string
              description: The policy.yaml/talos.yaml pattern that flagged this file, if any.
...
    Diff:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/GitChangeSummary"
        diff:
          type: string
          description: Unified diff text, or null if the run hasn't reached REVIEWING yet.
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, CHANGES_REQUESTED, EXPIRED]
    Approval:
      type: object
      required: [id, taskId, runId, approvalType, requestedAction, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        approvalType:
          type: string
        requestedAction:
          type: string
        status:
          $ref: "#/components/schemas/ApprovalStatus"
        requestedBy:
          type: string
          format: uuid
        approvedBy:
          type: string
          format: uuid
        approvedAt:
          type: string
          format: date-time
        notes:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        environment:
          type: string
          description: Phase 10 -- set only on DEPLOY-type approvals; which environment this approval gates.
...
    PullRequestStatus:
      type: string
      enum: [OPEN, MERGED, CLOSED]
    PullRequest:
      type: object
      required: [id, runId, provider, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        provider:
          type: string
        prNumber:
          type: integer
        url:
          type: string
        status:
          $ref: "#/components/schemas/PullRequestStatus"
        createdAt:
          type: string
          format: date-time
...
    InternalPullRequestRequest:
      type: object
      required: [branchName, commitSha]
      properties:
        branchName:
          type: string
        commitSha:
          type: string
    DeployStatus:
      type: string
      enum: [RUNNING, SUCCEEDED, FAILED]
    ProjectEnvironment:
      type: object
      description: "Phase 10 -- a project's deploy target, synced from talos.yaml's deploy: block each time a deploy is requested."
      required: [id, projectId, environment, provider, appId, approvalRequired]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        environment:
          type: string
        provider:
          type: string
        appId:
          type: string
        approvalRequired:
          type: boolean
        lastDeployStatus:
          $ref: "#/components/schemas/DeployStatus"
        lastDeployedAt:
          type: string
          format: date-time
        lastRunId:
          type: string
          format: uuid
    DeployTriggerResponse:
      type: object
      description: POST /runs/{id}/deploy's response -- discriminated on approvalRequired; approval is present only when true.
      required: [approvalRequired, environment]
      properties:
        approvalRequired:
          type: boolean
        approval:
          $ref: "#/components/schemas/Approval"
        environment:
          $ref: "#/components/schemas/ProjectEnvironment"